fix(routeButton): rename routeButton to backButton.
