+++ airsen-frontend/src/app/features/forum/threads/add-thread/add-thread.component.html
@@ -1,5 +1,5 @@
 <div class="page-container">
-    <route-button route="/forum" textContent="Retour au forum"></route-button>
+    <back-button route="/forum" textContent="Retour au forum"></back-button>
     <h2>Créer une nouvelle discussion</h2>
     @if (error$ != '') {
         <p class="error">{{ error$ }}</p>
