+++ airsen-frontend/src/app/features/forum/threads/threads-details/thread-details.component.html
@@ -1,5 +1,5 @@
 <div class="thread-container">
-  <route-button route="/forum" textContent="Retour au forum"></route-button>
+  <back-button route="/forum" textContent="Retour au forum"></back-button>
   <div class="thread-content">
     @if (thread$ | async) {
       <div class="thread-infos">
