+++ airsen-frontend/src/app/shared/components/backButton/back-button.component.scss
